improve: SQL을 queries.py로 분리
